wip: create helper for auth
